refactor: 로또 구매 금액과 티켓 수량의 자료형을 long으로 변경

+++ src/main/java/lotto/domain/Dealer.java
@@ -23,8 +23,8 @@ public class Dealer {
         return result;
     }
 
-    public float calculateEarningRate(int purchaseAmount) {
-        return earning / (float)purchaseAmount * 100;
+    public float calculateEarningRate(long purchaseAmount) {
+        return (float) (earning / purchaseAmount * 100.0);
     }
 
     private float calculateEarning() {
